legacy/code/requirements.txt
RPi.bme280
adafruit-circuitpython-ads1x15
flask